refactor: split AppStatus into components fix: change top-level options buttons order
