+++ src/react/AppStatus.tsx
@@ -1,31 +1,11 @@
 import { useEffect, useState } from 'react'
-import { proxy, useSnapshot } from 'valtio'
-import { activeModalStacks, hideModal, insertActiveModalStack, miscUiState } from '../globalState'
 import { guessProblem } from '../guessProblem'
-import { fsState } from '../loadSave'
-import { resetLocalStorageWorld } from '../browserfs'
 import styles from './loadingErrorScreen.module.css'
 import Button from './Button'
 import Screen from './Screen'
-import DiveTransition from './DiveTransition'
-import { isModalActive, useDidUpdateEffect } from './utils'
 
-const initialState = {
-  status: '',
-  lastStatus: '',
-  maybeRecoverable: true,
-  isError: false,
-  hideDots: false,
-}
-export const appStatusState = proxy(initialState)
-const resetState = () => {
-  Object.assign(appStatusState, initialState)
-}
-
-export default () => {
-  const { isError, lastStatus, maybeRecoverable, status, hideDots } = useSnapshot(appStatusState)
+export default ({ status, isError, hideDots = false, lastStatus = '', backAction = undefined, actionsSlot = undefined }) => {
   const [loadingDots, setLoadingDots] = useState('')
-  const isOpen = isModalActive('app-status')
 
   useEffect(() => {
     void statusRunner()
@@ -47,58 +27,25 @@ export default () => {
     void load()
   }
 
-  useDidUpdateEffect(() => {
-    // todo play effect only when world successfully loaded
-    if (!isOpen) {
-      const divingElem: HTMLElement = document.querySelector('#viewer-canvas')
-      divingElem.style.animationName = 'dive-animation'
-      divingElem.parentElement.style.perspective = '1200px'
-      divingElem.onanimationend = () => {
-        divingElem.parentElement.style.perspective = ''
-        divingElem.onanimationend = null
-      }
-    }
-  }, [isOpen])
-
   return (
-    <DiveTransition open={isOpen}>
-      <Screen
-        title={
-          <>
-            {status}
-            {isError || hideDots ? '' : loadingDots}
-            <p className={styles['potential-problem']}>{isError ? guessProblem(status) : ''}</p>
-            <p className={styles['last-status']}>{lastStatus ? `Last status: ${lastStatus}` : lastStatus}</p>
-          </>
-        }
-        backdrop='dirt'
-      >
-        {isError && (
-          <>
-            <Button hidden={!maybeRecoverable} label="Back" onClick={() => {
-              appStatusState.isError = false
-              resetState()
-              miscUiState.gameLoaded = false
-              miscUiState.loadedDataVersion = undefined
-              window.loadedData = undefined
-              if (activeModalStacks['main-menu']) {
-                insertActiveModalStack('main-menu')
-              } else {
-                hideModal(undefined, undefined, { force: true })
-              }
-            }}
-            ></Button>
-            <Button hidden={!(miscUiState.singleplayer && fsState.inMemorySave)} label="Reset world" onClick={() => {
-              if (window.confirm('Are you sure you want to delete all local world content?')) {
-                resetLocalStorageWorld()
-                window.location.reload()
-              }
-            }}
-            ></Button>
-            <Button onClick={() => window.location.reload()} label="Full Reload" ></Button>
-          </>
-        )}
-      </Screen>
-    </DiveTransition>
+    <Screen
+      title={
+        <>
+          {status}
+          {isError || hideDots ? '' : loadingDots}
+          <p className={styles['potential-problem']}>{isError ? guessProblem(status) : ''}</p>
+          <p className={styles['last-status']}>{lastStatus ? `Last status: ${lastStatus}` : lastStatus}</p>
+        </>
+      }
+      backdrop='dirt'
+    >
+      {isError && (
+        <>
+          {backAction && <Button label="Back" onClick={backAction} />}
+          {actionsSlot}
+          <Button onClick={() => window.location.reload()} label="Full Reload" ></Button>
+        </>
+      )}
+    </Screen>
   )
 }
